+++ src/living/helpers/gui2d/control.test.ts
@@ -1,5 +1,5 @@
 import { Control2D } from './control';
-import { describe, jest, it, expect } from '@jest/globals';
+import { describe, jest, it, expect, beforeAll } from '@jest/globals';
 import DOMMatrixImpl from '../../geometry/DOMMatrix';
 jest.mock('@jest/globals');
 
@@ -38,5 +38,71 @@ describe('Control', () => {
     control._updateTransform();
     
     expect((control as any)._renderingContext.getTransform()).toEqual(matrix);
+  });
+});
+describe('_calculateTransformMatrix', () => {
+  it('should calculate translateX correctly', () => {
+    const matrix = new DOMMatrixImpl([
+      1, 0, 0, 0,
+      0, 1, 0, 0,
+      0, 0, 1, 0,
+      0, 0, 0, 1
+    ]);
+    const control = new MockControl(matrix);
+    const transforms = [{ type: 'translateX', value: '10', unit: 'px' }];
+    const result = control['_calculateTransformMatrix'](transforms);
+    const expectedMatrix = new DOMMatrixImpl([
+      1, 0, 0, 0,  
+      0, 1, 0, 0,  
+      0, 0, 1, 0,  
+      10, 0, 0, 1
+    ]);
+
+    expect(result).toEqual(expectedMatrix);
+  });
+
+  it('should calculate rotate correctly', () => {
+    const matrix = new DOMMatrixImpl([
+      1, 0, 0, 0,
+      0, 1, 0, 0,
+      0, 0, 1, 0,
+      0, 0, 0, 1
+    ]);
+    const control = new MockControl(matrix);
+    const transforms = [{ type: 'rotate', value: '45', unit: 'deg' }];
+    const result = control['_calculateTransformMatrix'](transforms);
+    const cosValue = Number(Math.cos(45 * Math.PI / 180).toFixed(2));
+    const sinValue = Number(Math.sin(45 * Math.PI / 180).toFixed(2));
+    const expectedMatrix = new DOMMatrixImpl([
+      cosValue, sinValue, 0, 0,  
+      -sinValue, cosValue, 0, 0,  
+      0, 0, 1, 0,   
+      0, 0, 0, 1
+    ]);
+
+    expect(result).toEqual(expectedMatrix);
+  });
+
+  it('should calculate multiple transforms correctly', () => {
+    const matrix = new DOMMatrixImpl([
+      1, 0, 0, 0,
+      0, 1, 0, 0,
+      0, 0, 1, 0,
+      0, 0, 0, 1
+    ]);
+    const control = new MockControl(matrix);
+    const transforms = [
+      { type: 'translateX', value: '10', unit: 'px' },
+      { type: 'rotate', value: '90', unit: 'deg' },
+    ];
+    const result = control['_calculateTransformMatrix'](transforms);
+    const expectedMatrix = new DOMMatrixImpl([
+      0, 1, 0, 0,
+      -1, 0, 0, 0,
+      0, 0, 1, 0,
+      10, 0, 0, 1
+    ]);
+
+    expect(result).toEqual(expectedMatrix);
   });
 });
