+++ src/living/helpers/gui2d/control.ts
@@ -13,7 +13,7 @@ import { ShadowRootImpl } from '../../nodes/ShadowRoot';
 import { getInterfaceWrapper } from '../../../living/interfaces';
 import DOMMatrixImpl from '../../geometry/DOMMatrix';
 import { postMultiply } from '../matrix-functions';
-import { parseTransform, calculateTransformMatrix} from '../../cssom/parsers';
+import { parseTransform} from '../../cssom/parsers';
 
 type LengthPercentageDimension = string | number;
 type LayoutStyle = Partial<{
@@ -759,6 +759,41 @@ export class Control2D {
     renderingContext.drawImage(this._imageBitmap, rect.x, rect.y, rect.width, rect.height);
   }
   
+
+  private _calculateTransformMatrix(transforms: { type: string, value: string, unit: string }[]): DOMMatrixImpl {
+    let transformMatrix = new DOMMatrixImpl([
+      1, 0, 0, 0,   
+      0, 1, 0, 0,  
+      0, 0, 1, 0,   
+      0, 0, 0, 1
+    ]);
+    transforms.forEach(transform => {
+      if (transform.type === 'translateX') {
+        const x = parseFloat(transform.value);
+        const translateMatrix = new DOMMatrixImpl([
+          1, 0, 0, 0,  
+          0, 1, 0, 0,  
+          0, 0, 1, 0,  
+          x, 0, 0, 1
+        ]);
+        transformMatrix = postMultiply(transformMatrix, translateMatrix) as DOMMatrixImpl;
+      }
+      if (transform.type === 'rotate') {
+        const angle = parseFloat(transform.value);
+        const cosValue = Number(Math.cos(angle * Math.PI / 180).toFixed(2));
+        const sinValue = Number(Math.sin(angle * Math.PI / 180).toFixed(2));
+        const rotateMatrix = new DOMMatrixImpl([
+          cosValue, sinValue, 0, 0,  
+          -sinValue, cosValue, 0, 0,  
+          0, 0, 1, 0,   
+          0, 0, 0, 1
+        ]);
+        transformMatrix = postMultiply(transformMatrix, rotateMatrix) as DOMMatrixImpl;
+      }
+    });
+    return transformMatrix;
+  }
+
   private _updateCurrentTransformMatrix() {
     const element = this._element;
     const style = this._style;
@@ -768,7 +803,7 @@ export class Control2D {
     const transformStr = style.transform;
     const parentElement = element.parentElement;
     const transforms = parseTransform(transformStr)
-    this.currentTransformMatrix = calculateTransformMatrix(transforms);
+    this.currentTransformMatrix = this._calculateTransformMatrix(transforms);
     if (parentElement === null) {
       return;
     } else {
